# SavePoint/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SavePoint API",
    "version": "1.0.0",
    "description": "API REST do SavePoint: catálogo de jogos, comparação de preços entre lojas, histórico de preços, avaliações de usuários e lista de desejos."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/api/avaliacoes/usuario/{usuarioId}": {
      "get": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Listar avaliações de um usuário",
        "operationId": "avaliacaoUsuario_buscarPorUsuario",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de avaliações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvaliacaoUsuario"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Excluir todas as avaliações de um usuário",
        "operationId": "avaliacaoUsuario_deletarPorUsuario",
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso (sem conteúdo)"
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Listar todos os usuários",
        "operationId": "usuario_listarTodos",
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Cadastrar usuário",
        "operationId": "usuario_criar",
        "description": "A senha deve ter no mínimo 6 caracteres e nunca é devolvida nas respostas.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              },
              "example": {
                "nome": "Jean Silva",
                "email": "jean@exemplo.com",
                "senha": "senha123",
                "telefone": "47999990000"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou e-mail já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Buscar usuário por ID",
        "operationId": "usuario_buscarPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuários"
        ],
        "summary": "Atualizar usuário",
        "operationId": "usuario_atualizar",
        "description": "Atualiza nome, e-mail e telefone. Campos não informados são mantidos.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              },
              "example": {
                "nome": "Jean Silva",
                "email": "jean@exemplo.com",
                "senha": "senha123",
                "telefone": "47999990000"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "E-mail inválido ou já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuários"
        ],
        "summary": "Excluir usuário",
        "operationId": "usuario_deletar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Removido com sucesso (sem conteúdo)"
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/email/{email}": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Buscar usuário por e-mail",
        "operationId": "usuario_buscarPorEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{id}/senha": {
      "patch": {
        "tags": [
          "Usuários"
        ],
        "summary": "Atualizar senha do usuário",
        "operationId": "usuario_atualizarSenha",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarSenhaRequest"
              },
              "example": {
                "novaSenha": "novaSenha456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Senha inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{id}/ativar": {
      "patch": {
        "tags": [
          "Usuários"
        ],
        "summary": "Ativar usuário",
        "operationId": "usuario_ativar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário ativado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios/{id}/desativar": {
      "patch": {
        "tags": [
          "Usuários"
        ],
        "summary": "Desativar usuário",
        "operationId": "usuario_desativar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário desativado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErroResposta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Jogo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "dataLancamento": {
            "type": "string",
            "format": "date"
          },
          "imagemCapa": {
            "type": "string"
          },
          "classificacaoIndicativa": {
            "type": "string"
          },
          "desenvolvedora": {
            "type": "string"
          },
          "genero": {
            "type": "string"
          },
          "requisitosMinimos": {
            "$ref": "#/components/schemas/RequisitosMinimos"
          },
          "resumoAvaliacao": {
            "$ref": "#/components/schemas/ResumoAvaliacao"
          },
          "ofertas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaJogo"
            }
          }
        }
      },
      "RequisitosMinimos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "processador": {
            "type": "string"
          },
          "memoria": {
            "type": "string"
          },
          "placaDeVideo": {
            "type": "string"
          },
          "sistemaOperacional": {
            "type": "string"
          }
        }
      },
      "ResumoAvaliacao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "notaMedia": {
            "type": "number",
            "format": "double"
          },
          "totalAvaliacoes": {
            "type": "integer",
            "format": "int32"
          },
          "percentualRecomendacao": {
            "type": "number",
            "format": "double"
          },
          "principaisElogios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "principaisCriticas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "resumoGeradoIA": {
            "type": "string"
          },
          "dataGeracao": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Loja": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "urlLoja": {
            "type": "string"
          },
          "urlLogo": {
            "type": "string"
          }
        }
      },
      "OfertaJogo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "precoOriginal": {
            "type": "number",
            "format": "double"
          },
          "precoAtual": {
            "type": "number",
            "format": "double"
          },
          "percentualDesconto": {
            "type": "number",
            "format": "float"
          },
          "loja": {
            "$ref": "#/components/schemas/Loja"
          },
          "historico": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Historico"
            }
          }
        }
      },
      "Historico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "preco": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string",
            "writeOnly": true
          },
          "telefone": {
            "type": "string"
          },
          "dataCadastro": {
            "type": "string",
            "format": "date-time"
          },
          "ultimoAcesso": {
            "type": "string",
            "format": "date-time"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
...
      "AvaliacaoUsuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nota": {
            "type": "integer",
            "format": "int32"
          },
          "textoAvaliacao": {
            "type": "string"
          },
          "dataPublicacao": {
            "type": "string",
            "format": "date-time"
          },
          "curtidas": {
            "type": "integer",
            "format": "int32"
          },
          "horasJogadas": {
            "type": "integer",
            "format": "int32"
          },
          "indiceRecomendacao": {
            "type": "boolean"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "jogo": {
            "$ref": "#/components/schemas/Jogo"
          }
        }
      },
      "AtualizarSenhaRequest": {
        "type": "object",
        "properties": {
          "novaSenha": {
            "type": "string"
          }
        }
      },
      "ErroResposta": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "erro": {
            "type": "string"
          },
          "mensagem": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}